alfredxml: fix UnmarshalJSON with uninitialized Alfred2
The documentation suggests to decode monitoring data like this:

```
    ...
    var a alfredxml.Alfred2
    if err := dec.Decode(&a); err == io.EOF {
    ...
```

Doing it like this leaves `a` as an uninitialized "nil"-map. In
Alfred2.UnmarshalJSON a copy of `a` is made to convert to a plain type.
This avoids recursion in the json decoder.

```
    m := map[string]Data(*a)
```

Since a is not initialized, this will copy a nil map, thus no real
connection between `m` and `a` exists.

This patch fixes the issue by converting the pointer type. The pointer
will still point to `a`, whether it is initialized or not.

The Alfred.Unmarshal was changed to mirror the behaviour.